requirements.txt
arabic-reshaper==2.1.1
astroid==2.4.2
autopep8==1.5.4
certifi==2018.8.24
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cx-Freeze==6.4.2
defusedxml==0.7.0rc1
Django==2.2.7
django-autocomplete-light==3.8.1
django-ckeditor==6.0.0
django-debug-toolbar==1.9.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.2
easy-thumbnails==2.7.1
future==0.18.2
html5lib==1.1
idna==2.6
importlib-metadata==3.3.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.10
jsonfield==3.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.1
psycopg2-binary==2.8.6
py2exe==0.9.2.2
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2019.2
reportlab==3.5.56
requests==2.18.4
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
swapper==1.1.2.post1
toml==0.10.2
treepoem==3.7.0
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
zipp==3.4.0
